+++ assets/js/common/SuseMultiLinuxManagerClearSettingsDialog/SuseMultiLinuxManagerClearSettingsModal.test.jsx
