+++ assets/js/common/SuseMultiLinuxManagerConfig/SuseMultiLinuxManagerConfig.test.jsx
