+++ assets/js/common/SuseMultiLinuxManagerConfig/index.js
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: SUSE LLC
 // SPDX-License-Identifier: Apache-2.0
 
-import SuseMultiLinuxManagerConfig from './SuseManagerConfig';
+import SuseMultiLinuxManagerConfig from './SuseMultiLinuxManagerConfig';
 
 export default SuseMultiLinuxManagerConfig;
