+++ assets/js/common/SuseMultiLinuxManagerSettingsDialog/index.js
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: SUSE LLC
 // SPDX-License-Identifier: Apache-2.0
 
-import SuseMultiLinuxManagerSettingsModal from './SuseManagerSettingsModal';
+import SuseMultiLinuxManagerSettingsModal from './SuseMultiLinuxManagerSettingsModal';
 
 export default SuseMultiLinuxManagerSettingsModal;
