+++ assets/js/lib/api/suseMultiLinuxManagerSettings.js
