+++ assets/js/lib/model/suse_multilinux_manager.js
