+++ assets/js/pages/HostDetailsPage/suseMultiLinuxManagerMessaging.js
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: SUSE LLC
 // SPDX-License-Identifier: Apache-2.0
 
-import { SMLM_PRODUCT_LABEL } from '@lib/model/suse_manager';
+import { SMLM_PRODUCT_LABEL } from '@lib/model/suse_multilinux_manager';
 
 export const getSoftwareUpdatesErrorMessage = (errors) => {
   const hostNotFoundInSMLM = errors.some(
