+++ test/e2e/cypress-split-timings.json
@@ -77,7 +77,7 @@
       "duration": 0
     },
     {
-      "spec": "cypress/e2e/suse_manager_overviews.cy.js",
+      "spec": "cypress/e2e/suse_multimanager_overviews.cy.js",
       "duration": 40000
     },
     {
