+++ test/e2e/cypress/e2e/suse_multilinux_manager_overviews.cy.js
