+++ test/e2e/cypress/fixtures/smlm_credentials/certificates.js
